requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
asgiref==3.4.1
async-timeout==4.0.2
attrs==21.2.0
azure-common==1.1.27
azure-core==1.21.1
azure-identity==1.7.1
azure-mgmt-compute==23.1.0
azure-mgmt-core==1.3.0
azure-mgmt-network==19.3.0
azure-mgmt-resource==20.0.0
azure-monitor-query==1.0.1
backports.zoneinfo==0.2.1
black==21.12b0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.9
click==8.0.3
cryptography==36.0.1
Django==4.0
django-cors-headers==3.10.1
frozenlist==1.2.0
h11==0.12.0
httptools==0.3.0
idna==3.3
isodate==0.6.1
msal==1.16.0
msal-extensions==0.3.1
msrest==0.6.21
multidict==5.2.0
mypy-extensions==0.4.3
oauthlib==3.1.1
pathspec==0.9.0
platformdirs==2.4.0
portalocker==2.3.2
psycopg2-binary==2.9.2
pycparser==2.21
PyJWT==2.3.0
python-dotenv==0.19.2
PyYAML==6.0
requests==2.26.0
requests-oauthlib==1.3.0
six==1.16.0
sqlparse==0.4.2
tomli==1.2.3
typing-extensions==4.0.1
urllib3==1.26.7
uvicorn==0.16.0
uvloop==0.16.0
watchgod==0.7
websockets==10.1
yarl==1.7.2
